Update test of text with `position:absolute`
The following code:

```
<span class="show-for-sr">You're on page</span> 1
```

Will generate an element with `position: absolute`. When reading the
contents, it's not clear whether this element is supposed to be in the
 same paragraph or in a different one. Currently Capybara treats it as
if it were part of a different paragraph.

Since this could be the way screen readers read the text, I'm changing
the test to reflect this fact. We might change our minds in the future.